requirements.txt
anyio==3.4.0
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.4.0
bcrypt==3.2.0
bidict==0.21.4
CacheControl==0.12.10
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
DateTime==4.3
distlib==0.3.4
ecdsa==0.17.0
Faker==12.0.0
fastapi==0.70.1
fastapi-socketio==0.0.9
filelock==3.4.2
firebase-admin==5.2.0
google-api-core==2.4.0
google-api-python-client==2.36.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
greenlet==1.1.2
grpcio==1.43.0
grpcio-status==1.43.0
h11==0.12.0
httplib2==0.20.2
httptools==0.3.0
idna==3.3
importlib-metadata==4.10.0
iniconfig==1.1.1
msgpack==1.0.3
packaging==21.3
passlib==1.7.4
pip==21.3.1
pipenv==2022.1.8
platformdirs==2.4.1
pluggy==1.0.0
proto-plus==1.19.8
protobuf==3.19.3
psycopg2==2.9.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
PyJWT==2.3.0
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
python-engineio==4.3.0
python-jose==3.3.0
python-multipart==0.0.5
python-socketio==5.5.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
rsa==4.8
setuptools==59.4.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.29
starlette==0.16.0
toml==0.10.2
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
uvicorn==0.16.0
virtualenv==20.13.0
virtualenv-clone==0.5.7
watchgod==0.7
websockets==10.1
wheel==0.37.0
zipp==3.6.0
zope.interface==5.4.0